requirements.txt
anndata>=0.11.4
faiss-cpu>=1.12.0
matplotlib>=3.10.5
networkx>=3.3
numpy>=1.26.4
pandas>=2.3.1
pillow>=11.2.1
psutil>=7.0.0
scanpy>=1.10.4
scikit-learn>=1.7.1
scipy>=1.15.3
torch==2.1.0+cu121
torch_geometric>=2.3.1
torch-sparse==0.6.18+pt21cu121
torchvision==0.16.0+cu121
tqdm>=4.67.1
typing_extensions>=4.14.1
